chrono: convert to double precision for overflow purposes
# Conflicts:
#	ajek-framework/h/x-chrono.h

+++ ajek-framework/h/x-chrono.h
@@ -7,7 +7,7 @@ typedef u64 host_tick_t;
 struct HostClockTick {
 
 	static host_tick_t	s_ticks_per_second;
-	static float		s_ticks_per_second_f;
+	static double		s_ticks_per_second_f;
 
 	static void			Init();
 
@@ -20,11 +20,11 @@ struct HostClockTick {
 	}
 
 	__ai	auto		asTicks				() const	{ return m_val; }
-	__ai	float		asMicroseconds		() const	{ return (m_val * 1000000.0f)				/ s_ticks_per_second_f;	}
-	__ai	float		asMilliseconds		() const	{ return (m_val *    1000.0f)				/ s_ticks_per_second_f;	}
-	__ai	float		asSeconds			() const	{ return (m_val *       1.0f)				/ s_ticks_per_second_f;	}
-	__ai	float		asMinutes			() const	{ return (m_val * (1.0f / (60.0f      )))	/ s_ticks_per_second_f;	}
-	__ai	float		asHours				() const	{ return (m_val * (1.0f / (60.0f*60.0f)))	/ s_ticks_per_second_f;	}
+	__ai	double		asMicroseconds		() const	{ return (double(m_val) * 1000000.0)			/ s_ticks_per_second_f;	}
+	__ai	double		asMilliseconds		() const	{ return (double(m_val) *    1000.0)			/ s_ticks_per_second_f;	}
+	__ai	double		asSeconds			() const	{ return (double(m_val) *       1.0)			/ s_ticks_per_second_f;	}
+	__ai	double		asMinutes			() const	{ return (double(m_val) * (1.0 / (60.0      )))	/ s_ticks_per_second_f;	}
+	__ai	double		asHours				() const	{ return (double(m_val) * (1.0 / (60.0*60.0 )))	/ s_ticks_per_second_f;	}
 };
 
 // For getting system date information it should be preferable to use std::chrono, which has frameworks
